Fix authorization class to reprompt if the creds. in the keyring suck
This handles password changes ok, but won't handle username changes that well,
since there will now be two entries for the key in the keyring, and it will only
ever try with one of them, nor does it erase it when it doesn't work. This can
be fixed at a later date, because it doesn't really come up very often.