Revert "Synchronise some changes from the upstream"
This reverts commit 481182febd0e177c8ba5905e6194002f25046501.

With all due respect, this is a semi-random combination of copy/paste.
As-is the branch is in foobar'd state and this patch doesn't help et al.

Converging with upstream get my +1000, although let's first have a
branch in usable state.

The follow-up commits do that, so let's revert this for now.