gtimelog: go through all task data via the search box
Unlike upstream, we have a segmented task list. In consists of four
components - customer:project:group:task

Currently when the user writes "test", while searching for tasks, the
text is matched only with the "group" entries. This is fairly misleading
since the user may be looking for the any of the four.

Reported-by: Eze Garcia <ezequiel@collabora.com>
Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>