gtimelog: display the full tree on match
Earlier commit started matching any of the task data, although it did
not show the full task tree, when a match was made. Namely for a given:

customer:project:group:task

If we matched the group, none of the tasks were listed. This commit
fixes that adding a blucky comment explaining the current, IMHO weird,
state of affairs.

Fixes: 11bdaae ("gtimelog: go through all task data via the search box")
Reported-by: Fadi Abou Younes <fadi.abouyounes@collabora.com>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>