group reports by task
The entries are hierarchical and we expect at least 2 : separators.
The comment is after the last separator.
Before the last separator is the task id.
Before the 2nd to last separator is the category.

This sort of takes care of the optional per-user top-level,
although it's currently included in the category.

(note: this function is used for weekly/monthly reports but not daily
ones, which don't matter to us)