main window: add summary for the last month as well
Use-case: I know I worked 20 days last month. I forgot to record a few
hours on Friday. If I fix this on Sunday, then I can just edit the
timelog and see the monthly stat and verify that the 20 days total
matches the expected value. Then submit it to chronophage.

However, I can't do the same on Monday, because we're already in March.
Fix this problem by providing stats for both the current month and the
last month.